fix: widen ledger sequence to u64 in charge anti-frontrunning salt
env.ledger().sequence() returns u32 in soroban-sdk 22, but the salt buffer
reserves 8 bytes for it, so copy_from_slice panicked with a 4-vs-8 byte
mismatch on every interval charge. Widen to u64 to fix the host panic that
broke auto-pause, grace buyout, fee routing, scheduled cancel, split billing
and subscription-transfer charges.

+++ contracts/subscription_vault/src/subscription.rs
@@ -1044,7 +1044,7 @@ pub fn do_grace_buyout(
                 let mut salt_buf = [0u8; 20];
                 salt_buf[..4].copy_from_slice(&subscription_id.to_be_bytes());
                 salt_buf[4..12].copy_from_slice(&sub.last_payment_timestamp.to_be_bytes());
-                salt_buf[12..20].copy_from_slice(&env.ledger().sequence().to_be_bytes());
+                salt_buf[12..20].copy_from_slice(&(env.ledger().sequence() as u64).to_be_bytes());
                 let salt_input = soroban_sdk::Bytes::from_slice(env, &salt_buf);
                 env.crypto().sha256(&salt_input).into()
             },
